fix: recurse into nested #if chains inside a reachable arm
_strip_inactive_if_zero_blocks() used a flat depth counter for any #if
nested inside a live (unmasked) arm -- it tracked depth well enough to
find the matching #endif, but never evaluated the nested condition
itself, so a dead nested #if 0 compatibility stub inside an otherwise-live
arm still leaked its content into the shadow/requirements scan. A genuine
active C++20 declaration elsewhere in the header could then be wrongly
shadowed by the dead nested stub, making the detector under-report and
skip -std=gnu++20 for a header that genuinely needed it.

Rewritten around an explicit per-depth stack: each nesting level
independently recognizes zero/true/feature-guard conditions the same way
the top level does, and any level with an unreachable ancestor is
unconditionally unreachable regardless of its own condition. This also
naturally handles arbitrary nesting depth, not just one extra level.

Split the now-2000+-line cpp20 detector test file: preprocessor-guard-
masking regressions (type-name shadowing, elif/else reachability,
__cplusplus/feature-test-macro circularity, nested #if chains) move to
tests/test_dumper_ast_config_cpp20_pp_guards.py; the parent file keeps
literal/comment/string handling and abbreviated-auto-parameter detection.

Addresses a Codex review finding on PR #622.
