+++ changelog.d/20260724_002813_noreply_abicheck_toolchain_audit_q0caml.md
@@ -1,0 +1,79 @@
+<!--
+A new changelog fragment. See changelog.d/README.md for the workflow.
+
+Uncomment exactly ONE '### <Category>' section below (remove its comment
+wrapper) and replace the example bullet with your entry, written the way
+it should read in CHANGELOG.md. Delete the other sections.
+-->
+
+<!--
+### Added
+
+- **Short bold summary** — the rest of the sentence: what changed, for
+  whom, and why it matters. Backtick identifiers like `ChangeKind` or
+  `--policy-file`.
+
+-->
+<!--
+### Changed
+
+- **Short bold summary** — the rest of the sentence: what changed, for
+  whom, and why it matters. Backtick identifiers like `ChangeKind` or
+  `--policy-file`.
+
+-->
+<!--
+### Deprecated
+
+- **Short bold summary** — the rest of the sentence: what changed, for
+  whom, and why it matters. Backtick identifiers like `ChangeKind` or
+  `--policy-file`.
+
+-->
+<!--
+### Removed
+
+- **Short bold summary** — the rest of the sentence: what changed, for
+  whom, and why it matters. Backtick identifiers like `ChangeKind` or
+  `--policy-file`.
+
+-->
+### Fixed
+
+- **C++20 detection now recursively evaluates nested `#if` chains inside
+  a reachable arm, instead of blindly copying them through.** The
+  preprocessor-guard masker used a flat depth counter for any `#if`
+  nested inside a live arm — it tracked depth well enough to find the
+  matching `#endif`, but never evaluated the nested condition itself. A
+  dead nested `#if 0` compatibility stub inside an otherwise-live arm
+  could therefore still shadow a genuine, active C++20 declaration
+  elsewhere in the header (and, symmetrically, live-looking nested code
+  inside a dead ancestor arm was already handled correctly). Rewritten
+  as an explicit per-depth stack so every nesting level is masked with
+  the same rules as the top level, recursively.
+
+-->
+<!--
+### Performance
+
+- **Short bold summary** — the rest of the sentence: what changed, for
+  whom, and why it matters. Backtick identifiers like `ChangeKind` or
+  `--policy-file`.
+
+-->
+<!--
+### Security
+
+- **Short bold summary** — the rest of the sentence: what changed, for
+  whom, and why it matters. Backtick identifiers like `ChangeKind` or
+  `--policy-file`.
+
+-->
+<!--
+### Documentation
+
+- **Short bold summary** — the rest of the sentence: what changed, for
+  whom, and why it matters. Backtick identifiers like `ChangeKind` or
+  `--policy-file`.
+
+-->
